Add test skeleton with jest and chai
