+++ package.json
@@ -7,7 +7,7 @@
   "private": true,
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node_modules/jest/bin/jest.js test/*.test.js",
     "start": "./node_modules/.bin/babel-node server.js",
     "build": "webpack --config webpack.config.js",
     "startmon": "nodemon server.js",
@@ -50,7 +50,9 @@
     "lint-staged": "^4.0.3",
     "nodemon": "^1.14.11",
     "pre-commit": "^1.2.2",
-    "prettier": "^1.13.4"
+    "prettier": "^1.13.4",
+    "supertest": "^3.1.0",
+    "chai": "^4.1.2"
   },
   "dependencies": {
     "@warren-bank/ethereumjs-tx-unsign": "^2.0.0",
@@ -66,6 +68,7 @@
     "ethereumjs-wallet": "^0.6.0",
     "express": "^4.16.3",
     "express-prettify": "0.0.10",
+    "http-status-codes": "^1.3.0",
     "morgan": "^1.9.0",
     "solc": "^0.4.20",
     "swagger": "^0.7.5",
